requirements.txt
django==4.0.2
djangorestframework==3.13.1
django-cors-headers==3.10.1
django-environ==0.9.0